# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /proveedores:
    get:
      operationId: getProveedores
      tags: [proveedores]
      summary: Listar proveedores adjudicados
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: q
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Lista de proveedores
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedoresResponse"
  /proveedores/{ruc}:
    get:
      operationId: getProveedor
      tags: [proveedores]
      summary: Ficha de un proveedor
      parameters:
        - name: ruc
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Detalle del proveedor
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProveedorDetalle"
        "404":
          description: No encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /observatorio/rankings:
    get:
      operationId: getRankings
      tags: [observatorio]
      summary: Rankings de entidades y proveedores
      parameters:
        - name: anio
          in: query
          schema:
            type: ["integer", "null"]
        - name: ubigeo
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: Rankings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rankings"
  /observatorio/alertas:
    get:
      operationId: getAlertas
      tags: [observatorio]
      summary: Contrataciones con indicadores de alerta
      parameters:
        - name: ubigeo
          in: query
          schema:
            type: ["string", "null"]
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Lista de alertas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AlertaContratacion"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    Contratacion:
      type: object
      properties:
        ocid:
          type: string
        nomenclatura:
          type: ["string", "null"]
        titulo:
          type: string
        descripcion:
          type: ["string", "null"]
        tipo:
          type: ["string", "null"]
        procedimiento:
          type: ["string", "null"]
        estado:
          type: ["string", "null"]
        entidadRuc:
          type: ["string", "null"]
        entidadNombre:
          type: ["string", "null"]
        proveedorRuc:
          type: ["string", "null"]
        proveedorNombre:
          type: ["string", "null"]
        ubigeoCodigo:
          type: ["string", "null"]
        ubigeoDistrito:
          type: ["string", "null"]
        ubigeoProvincia:
          type: ["string", "null"]
        montoReferencial:
          type: ["number", "null"]
        montoAdjudicado:
          type: ["number", "null"]
        moneda:
          type: ["string", "null"]
        fechaConvocatoria:
          type: ["string", "null"]
        fechaAdjudicacion:
          type: ["string", "null"]
        fechaContrato:
          type: ["string", "null"]
        plazoEjecucionDias:
          type: ["integer", "null"]
        createdAt:
          type: ["string", "null"]
        updatedAt:
          type: ["string", "null"]
      required:
        - ocid
        - titulo
...
    UbigeoBasico:
      type: object
      properties:
        codigo:
          type: string
        departamento:
          type: string
        provincia:
          type: string
        distrito:
          type: string
        latitud:
          type: ["number", "null"]
        longitud:
          type: ["number", "null"]
      required:
        - codigo
        - departamento
        - provincia
        - distrito
...
    ProveedorBasico:
      type: object
      properties:
        ruc:
          type: string
        razonSocial:
          type: string
        ubigeoCodigo:
          type: ["string", "null"]
        vigenteRnp:
          type: ["string", "null"]
      required:
        - ruc
        - razonSocial
    ProveedorDetalle:
      allOf:
        - $ref: "#/components/schemas/ProveedorBasico"
        - type: object
          properties:
            ubigeo:
              $ref: "#/components/schemas/UbigeoBasico"
            totalAdjudicaciones:
              type: integer
            montoTotalGanado:
              type: ["number", "null"]
            entidadesUnicas:
              type: integer
            adjudicacionesRecientes:
              type: array
              items:
                $ref: "#/components/schemas/Contratacion"
            topEntidades:
              type: array
              items:
                $ref: "#/components/schemas/EntidadRanking"
            evolucionMensual:
              type: array
              items:
                $ref: "#/components/schemas/PuntoEvolucion"
          required:
            - totalAdjudicaciones
    ProveedoresResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProveedorBasico"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
      required:
        - data
        - total
        - page
        - limit
        - pages
    ProveedorRanking:
      type: object
      properties:
        ruc:
          type: string
        razonSocial:
          type: string
        totalAdjudicaciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
      required:
        - ruc
        - razonSocial
        - totalAdjudicaciones
    EntidadRanking:
      type: object
      properties:
        ruc:
          type: string
        nombre:
          type: string
        totalContrataciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
      required:
        - ruc
        - nombre
        - totalContrataciones
...
    PuntoEvolucion:
      type: object
      properties:
        mes:
          type: string
        anio:
          type: integer
        periodo:
          type: string
        totalContrataciones:
          type: integer
        montoTotal:
          type: ["number", "null"]
      required:
        - mes
        - anio
        - periodo
        - totalContrataciones
...
    Rankings:
      type: object
      properties:
        topEntidadesMonto:
          type: array
          items:
            $ref: "#/components/schemas/EntidadRanking"
        topProveedoresMonto:
          type: array
          items:
            $ref: "#/components/schemas/ProveedorRanking"
        entidadesMasDirectas:
          type: array
          items:
            $ref: "#/components/schemas/EntidadRanking"
      required:
        - topEntidadesMonto
        - topProveedoresMonto
        - entidadesMasDirectas
    AlertaContratacion:
      type: object
      properties:
        ocid:
          type: string
        titulo:
          type: string
        tipoAlerta:
          type: string
        descripcionAlerta:
          type: string
        montoReferencial:
          type: ["number", "null"]
        montoAdjudicado:
          type: ["number", "null"]
        entidadNombre:
          type: ["string", "null"]
        distrito:
          type: ["string", "null"]
        procedimiento:
          type: ["string", "null"]
      required:
        - ocid
        - titulo
        - tipoAlerta
        - descripcionAlerta
...