# lib/api-zod/src/generated/types/proveedorDetalle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
 * OpenAPI spec version: 0.1.0
 */
import type { Contratacion } from "./contratacion";
import type { EntidadRanking } from "./entidadRanking";
import type { ProveedorBasico } from "./proveedorBasico";
import type { PuntoEvolucion } from "./puntoEvolucion";
import type { UbigeoBasico } from "./ubigeoBasico";

export type ProveedorDetalle = ProveedorBasico & {
  ubigeo?: UbigeoBasico;
  totalAdjudicaciones: number;
  /** @nullable */
  montoTotalGanado?: number | null;
  entidadesUnicas?: number;
  adjudicacionesRecientes?: Contratacion[];
  topEntidades?: EntidadRanking[];
  evolucionMensual?: PuntoEvolucion[];
};