# lib/api-zod/src/generated/types/proveedoresResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
 * OpenAPI spec version: 0.1.0
 */
import type { ProveedorBasico } from "./proveedorBasico";

export interface ProveedoresResponse {
  data: ProveedorBasico[];
  total: number;
  page: number;
  limit: number;
  pages: number;
}