# lib/api-zod/src/generated/types/rankings.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vecino Vigilante - API de contrataciones públicas de Chanchamayo y Junín
 * OpenAPI spec version: 0.1.0
 */
import type { EntidadRanking } from "./entidadRanking";
import type { ProveedorRanking } from "./proveedorRanking";

export interface Rankings {
  topEntidadesMonto: EntidadRanking[];
  topProveedoresMonto: ProveedorRanking[];
  entidadesMasDirectas: EntidadRanking[];
}